Crate eip681

Parser for EIP-681 transaction requests.

The top level of the parsing tree is the type EthereumTransactionRequest.

ABNF Syntax

request          = schema_prefix target_address [ "@" chain_id ] [ "/" function_name ] [ "?" parameters ]
schema_prefix    = "ethereum" ":" [ "pay-" ]
target_address   = ethereum_address
chain_id         = 1*DIGIT
function_name    = STRING
ethereum_address = ( "0x" 40*HEXDIG ) / ENS_NAME
parameters       = parameter *( "&" parameter )
parameter        = key "=" value
key              = "value" / "gas" / "gasLimit" / "gasPrice" / TYPE
value            = number / ethereum_address / STRING
number           = [ "-" / "+" ] *DIGIT [ "." 1*DIGIT ] [ ( "e" / "E" ) [ 1*DIGIT ] ]

Where TYPE is a standard ABI type name, as defined in Ethereum Contract ABI specification. STRING is a URL-encoded unicode string of arbitrary length, where delimiters and the percentage symbol (%) are mandatorily hex-encoded with a % prefix.

For the syntax of ENS_NAME, please consult ERC-137 defining Ethereum Name Service.

See ABNF core rules for general information about the ABNF format.

Modules

error

Any and all errors encountered during parsing and validation.

Structs

Digits

Zero or more consecutive digits.

EnsName

Ethereum Name Service name.

EthereumAbiTypeName

An Ethereum ABI type name.

EthereumTransactionRequest

Ethereum transaction request.

HexDigits

Zero or more consecutive hexadecimal digits.

Number

A parsed number.

Parameters

A collection of Parameter.

SchemaPrefix

Schema prefix.

UrlEncodedUnicodeString

A URL-encoded unicode string of arbitrary length, where delimiters and the percentage symbol are mandatorily hex-encoded with a % prefix.

Enums

EthereumAddress

An Ethereum address.

Parameter

A key-value pair, where the the type of the value depends upon the key.

Value

A parameter value.